+++ pkg/cmd/aiassistantversion.go
@@ -108,6 +108,10 @@ var aiAssistantsVersionsUpdate = requestflag.WithInnerFlags(cli.Command{
 			Name:     "telephony-settings",
 			BodyPath: "telephony_settings",
 		},
+		&requestflag.Flag[[]string]{
+			Name:     "tool-id",
+			BodyPath: "tool_ids",
+		},
 		&requestflag.Flag[[]map[string]any]{
 			Name:     "tool",
 			Usage:    "The tools that the assistant can use. These may be templated with [dynamic variables](https://developers.telnyx.com/docs/inference/ai-assistants/dynamic-variables)",
@@ -227,7 +231,7 @@ var aiAssistantsVersionsUpdate = requestflag.WithInnerFlags(cli.Command{
 	"voice-settings": {
 		&requestflag.InnerFlag[string]{
 			Name:       "voice-settings.voice",
-			Usage:      "The voice to be used by the voice assistant. Check the full list of [available voices](https://developers.telnyx.com/api/call-control/list-text-to-speech-voices) via our voices API.\nTo use ElevenLabs, you must reference your ElevenLabs API key as an integration secret under the `api_key_ref` field. See [integration secrets documentation](https://developers.telnyx.com/api-reference/integration-secrets/create-a-secret) for details. For Telnyx voices, use `Telnyx.<model_id>.<voice_id>` (e.g. Telnyx.KokoroTTS.af_heart)",
+			Usage:      "The voice to be used by the voice assistant. Check the full list of [available voices](https://developers.telnyx.com/docs/tts-stt/tts-available-voices) via our voices API.\nTo use ElevenLabs, you must reference your ElevenLabs API key as an integration secret under the `api_key_ref` field. See [integration secrets documentation](https://developers.telnyx.com/api-reference/integration-secrets/create-a-secret) for details. For Telnyx voices, use `Telnyx.<model_id>.<voice_id>` (e.g. Telnyx.KokoroTTS.af_heart)",
 			InnerField: "voice",
 		},
 		&requestflag.InnerFlag[string]{