+++ pkg/cmd/whatsappbusinessaccountphonenumber.go
@@ -67,6 +67,7 @@ var whatsappBusinessAccountsPhoneNumbersInitializeVerification = cli.Command{
 		},
 		&requestflag.Flag[string]{
 			Name:     "verification-method",
+			Usage:    `Allowed values: "sms", "voice".`,
 			Default:  "sms",
 			BodyPath: "verification_method",
 		},