+++ pkg/cmd/whatsappmessagetemplate.go
@@ -40,10 +40,12 @@ var whatsappMessageTemplatesUpdate = cli.Command{
 		},
 		&requestflag.Flag[string]{
 			Name:     "category",
+			Usage:    `Allowed values: "MARKETING", "UTILITY", "AUTHENTICATION".`,
 			BodyPath: "category",
 		},
 		&requestflag.Flag[[]map[string]any]{
 			Name:     "component",
+			Usage:    "Updated template components. Same structure as the create request.",
 			BodyPath: "components",
 		},
 	},